Refactoring app DB usage - one DB load per request

+++ packages/server/src/api/controllers/table/internal.js
@@ -1,4 +1,3 @@
-const CouchDB = require("../../../db")
 const linkRows = require("../../../db/linkedRows")
 const { getRowParams, generateTableID } = require("../../../db/utils")
 const { FieldTypes } = require("../../../constants")
@@ -9,12 +8,13 @@ const {
   handleDataImport,
 } = require("./utils")
 const usageQuota = require("../../../utilities/usageQuota")
+const { getAppDB } = require("@budibase/backend-core/context")
+const env = require("../../../environment")
 const { cleanupAttachments } = require("../../../utilities/rowProcessor")
 const { runStaticFormulaChecks } = require("./bulkFormula")
 
 exports.save = async function (ctx) {
-  const appId = ctx.appId
-  const db = new CouchDB(appId)
+  const db = getAppDB()
   const { dataImport, ...rest } = ctx.request.body
   let tableToSave = {
     type: "table",
@@ -36,8 +36,7 @@ exports.save = async function (ctx) {
   // saving a table is a complex operation, involving many different steps, this
   // has been broken out into a utility to make it more obvious/easier to manipulate
   const tableSaveFunctions = new TableSaveFunctions({
-    db,
-    ctx,
+    user: ctx.user,
     oldTable,
     dataImport,
   })
@@ -82,7 +81,6 @@ exports.save = async function (ctx) {
   // update linked rows
   try {
     const linkResp = await linkRows.updateLinks({
-      appId,
       eventType: oldTable
         ? linkRows.EventType.TABLE_UPDATED
         : linkRows.EventType.TABLE_SAVE,
@@ -107,13 +105,12 @@ exports.save = async function (ctx) {
 
   tableToSave = await tableSaveFunctions.after(tableToSave)
   // has to run after, make sure it has _id
-  await runStaticFormulaChecks(appId, tableToSave, { oldTable })
+  await runStaticFormulaChecks(tableToSave, { oldTable })
   return tableToSave
 }
 
 exports.destroy = async function (ctx) {
-  const appId = ctx.appId
-  const db = new CouchDB(appId)
+  const db = getAppDB()
   const tableToDelete = await db.get(ctx.params.tableId)
 
   // Delete all rows for that table
@@ -127,7 +124,6 @@ exports.destroy = async function (ctx) {
 
   // update linked rows
   await linkRows.updateLinks({
-    appId,
     eventType: linkRows.EventType.TABLE_DELETE,
     table: tableToDelete,
   })
@@ -136,24 +132,25 @@ exports.destroy = async function (ctx) {
   await db.remove(tableToDelete)
 
   // remove table search index
-  const currentIndexes = await db.getIndexes()
-  const existingIndex = currentIndexes.indexes.find(
-    existing => existing.name === `search:${ctx.params.tableId}`
-  )
-  if (existingIndex) {
-    await db.deleteIndex(existingIndex)
+  if (!env.isTest()) {
+    const currentIndexes = await db.getIndexes()
+    const existingIndex = currentIndexes.indexes.find(
+      existing => existing.name === `search:${ctx.params.tableId}`
+    )
+    if (existingIndex) {
+      await db.deleteIndex(existingIndex)
+    }
   }
 
   // has to run after, make sure it has _id
-  await runStaticFormulaChecks(appId, tableToDelete, { deletion: true })
-  await cleanupAttachments(appId, tableToDelete, { rows })
+  await runStaticFormulaChecks(tableToDelete, { deletion: true })
+  await cleanupAttachments(tableToDelete, { rows })
   return tableToDelete
 }
 
 exports.bulkImport = async function (ctx) {
-  const appId = ctx.appId
-  const table = await getTable(appId, ctx.params.tableId)
+  const table = await getTable(ctx.params.tableId)
   const { dataImport } = ctx.request.body
-  await handleDataImport(appId, ctx.user, table, dataImport)
+  await handleDataImport(ctx.user, table, dataImport)
   return table
 }
